GermaParl Corpus Class.

The class inherits from the Corpus class defined in the package 'polmineR'. So far, it only adds a method 'summary'.

GermaParl

Format

Class 'R6ClassGenerator' <GermaParl> object generator
  Inherits from: <polmineR::Corpus>
  Public:
    initialize: function (corpus = "GERMAPARL", pAttribute = NULL) 
    summary: function (sAttribute) 
    clone: function (deep = FALSE) 
  Parent env: <environment: namespace:GermaParl>
  Locked objects: TRUE
  Locked class: FALSE
  Portable: TRUE 
 - attr(*, "name")= chr "GermaParl_generator"

Methods

summary(sAttribute)

Get number of tokens in subcorpora that are prepared on the basis of the s-attribute provided.

Examples

use("GermaParl")

#> ... resetting CORPUS_REGISTRY environment variable:

#>     /home/blaette/R/x86_64-pc-linux-gnu-library/3.4/GermaParl/extdata/cwb/registry

#> ... unloading rcqp library

#> ... reloading rcqp library

#> ... status: OK

GParl <- GermaParl$new()

#> Issue:
#> 	Data access error (CL: can't load and/or create necessary data)
#> Perhaps the corpus GERMAPARL is not accessible from the machine you are using.

#> Error in rcqp::cqi_attributes(corpus, type): Error: no such corpus

dt <- GParl$summary("year")

#> Error in eval(expr, envir, enclos): Objekt 'GParl' nicht gefunden

barplot(
  height = dt[["size"]], names.arg = dt[["year"]],
  xlab = "year", ylab = "number of tokens"
)

#> Error in dt[["size"]]: Objekt des Typs 'closure' ist nicht indizierbar