Table with information on GermaParl by year

A dataset with information on the corpus on a year-by-year basis is included in the package to be included in the data report of the package vignette. The code used to generate the data is reported in the examples section.

germaparl_by_year

Format

An object of class data.table (inherits from data.frame) with 22 rows and 6 columns.

Details

Note that the table is based on v1.0.5 of the corpus.

Examples

if (FALSE) {
dts <- lapply(
 13:17,
 function(lp){
  print(lp)
  P <- partition("GERMAPARL", lp = lp)
  dates <- as.Date(s_attributes(P, "date"))
  dt <- count(P, p_attribute = "lemma")
  unknown <- round(sum(dt[grepl("#unknown#", dt[["lemma"]])][["count"]]) / size(P), digits = 3)

  data.table(
    lp = lp,
    protocols = length(unique(s_attributes(P, "session"))),
    first = min(dates, na.rm = TRUE),
    last = min(dates, na.rm = TRUE),
    size = size(P),
    unknown = unknown
  )
}
)
germaparl_by_lp <- rbindlist(dts)
}