Unload and relaod GERMAPARL corpus

After adding attributes to the corpus, new annotations will not be available until the internal C representation of the corpus is deleted and reloaded. The function is used internally by functions adding s-attributes to the corpus. It is exported and documented as a matter of transparency.

germaparl_refresh(
  session_registry_dir = polmineR::registry(),
  system_registry_dir = getOption("polmineR.corpus_registry"),
  sample = FALSE
)

Arguments

session_registry_dir
The temporary session registry directory created and used by the polmineR package.

system_registry_dir
The non-temporary system registry directory.

sample
A logical value, if TRUE, the GERMAPARLSAMPLE corpus will be used. The default is FALSE, and the GERMAPARL corpus will be refreshed.

Details

The function ensures that a modified/updated registry file will be copied from the system registry directory (directory where registry files are stored permanently) to the temporary registry directory used by polmineR.