Use topicmodels prepared for GermaParl.

A set of LDA topicmodels is part of the Zenodo release of GermaParl (k between 100 and 450). These topic models can be downloaded using germaparl_download_lda and loaded using germaparl_load_topicmodel. Use germaparl_encode_lda_topics to add a structural attribute with topic information for speeches to the encoded corpus.

germaparl_download_lda(
  k = c(100L, 150L, 175L, 200L, 225L, 250L, 275L, 300L, 350L, 400L, 450L),
  doi = "https://doi.org/10.5281/zenodo.3742113",
  data_dir,
  sample = FALSE
)

germaparl_encode_lda_topics(
  k = 200,
  n = 5,
  registry_dir = cwb_registry_dir(),
  data_dir,
  sample = FALSE
)

germaparl_load_topicmodel(
  k,
  registry_dir = cwbtools::cwb_registry_dir(),
  verbose = TRUE,
  sample = FALSE
)

Arguments

k
A numeric or integer vector, the number of topics of the topicmodel. Multiple values can be provided to download several topic models at once.

doi
The DOI of GermaParl at Zenodo (stated as URL).

data_dir
The data directory with the binary files of the GERMAPARL corpus. If missing, the directory will be guessed using the function cwb::cwb_corpus_dir

sample
A logical value, if TRUE, use GERMAPARLSAMPLE corpus rather than GERMAPARL.

n
Number of topics to write to corpus

registry_dir
The registry directory where the registry file for GERMAPARL is located.

verbose
logical

Value

The functions germaparl_download_lda and germaparl_encode_lda_topics are returned for their side effects (downloading topic model and encoding topic model). They return TRUE if the operation has been succesful. The germaparl_download_lda function will return a LDA_Gibbs class object as defined in the topicmodels package.

Details

The function germaparl_download_lda will download an rds-file that will be stored in the data directory of the GermaParl corpus.

germaparl_encode_lda_topics will add a new s-attributes topics to GermaParl corpus with topicmodel for k topics. The n topics for speeches will be written to the corpus. A requirement for the function to work is that the s-attribute speech has been generated beforehand using germaparl_encode_speeches.

germaparl_load_topicmodel will load a topicmodel into memory. The function will return a LDA_Gibbs topicmodel, if the topicmodel for k is present; NULL if the topicmodel has not yet been downloaded.

See also