GermaParl R Data Package.

GermaParl is a corpus of parliamentary debates in the German Bundestag. The package offers a convenient dissemination mechanism for the GermaParl corpus. The corpus has been linguistically annotated and indexed using the data format of the Corpus Workbench (CWB). To make full use if this data format, working with GermaParl in combination with the polmineR package is recommended. After installation, only a small sample corpus will be included in the package ("GERMAPARLMINI"). To download the full corpus from the open science repository Zenodo, use the germaparl_download_corpus function. The package offers further functionality to augment the corpus (see the the examples section below).

Details

The GermaParl R package and the GermaParl corpus are two different pieces of research data: The package offers a mechanism to ship, easily install and augment the data. The indexed corpus is the actual data. Package and corpus have different version numbers and should be quoted in combination in publications. We recommend to follow the instructions you see when calling citation(package = "GermaParl"). To ensure that the recommended citation fits the corpus you use, the citation for the corpus is available only when a version of GermaParl has been downloaded and installed.

References

Blaette, Andreas (2018): "Using Data Packages to Ship Annotated Corpora of Parliamentary Protocols: The GermaParl R Package". ISBN 979-10-95546-02-3. Available online at http://lrec-conf.org/workshops/lrec2018/W2/pdf/15_W2.pdf.

See also

Examples

# This example uses the GERMAPARLSAMPLE corpus rather than the full GERMAPARL 
# so that the time consumed for testing the code is not excessive. To apply
# everything on GERMAPARL rather than GERMAPARLSAMPLE, set variable 'samplemode' 
# to FALSE, or simply omit argument 'sample'.

samplemode <- TRUE
corpus_id <- "GERMAPARLSAMPLE"
# corpus_id <- "GERMAPARL"

# This example assumes that the directories used by the CWB do not yet exist, so
# temporary directories are created.
cwb_dirs <- cwbtools::create_cwb_directories(prefix = tempdir(), ask = interactive())

#> Using directory '/var/folders/m_/431fjnbs1t32_62d35wvs7pr0000gp/T//RtmpAzVmz4' as parent directory for registry directory and the corpus directory. (Directory already exists.)

#> ... registry directory '/var/folders/m_/431fjnbs1t32_62d35wvs7pr0000gp/T//RtmpAzVmz4/registry' has been created successfully.

#> ... environment variable CORPUS_REGISTRY set as: /var/folders/m_/431fjnbs1t32_62d35wvs7pr0000gp/T//RtmpAzVmz4/registry

#> ... corpus directory '/var/folders/m_/431fjnbs1t32_62d35wvs7pr0000gp/T//RtmpAzVmz4/indexed_corpora' has been created successfully.

registry_tmp <- cwb_dirs[["registry_dir"]]

# Download corpus from Zenodo
germaparl_download_corpus(
  registry_dir = registry_tmp,
  corpus_dir = cwb_dirs[["corpus_dir"]],
  verbose = FALSE,
  sample = samplemode
)

#>   |                                                                              |                                                                      |   0%  |                                                                              |==                                                                    |   3%  |                                                                              |=====                                                                 |   7%  |                                                                              |=======                                                               |  10%  |                                                                              |=========                                                             |  13%  |                                                                              |============                                                          |  17%  |                                                                              |==============                                                        |  20%  |                                                                              |================                                                      |  23%  |                                                                              |===================                                                   |  27%  |                                                                              |=====================                                                 |  30%  |                                                                              |=======================                                               |  33%  |                                                                              |==========================                                            |  37%  |                                                                              |============================                                          |  40%  |                                                                              |==============================                                        |  43%  |                                                                              |=================================                                     |  47%  |                                                                              |===================================                                   |  50%  |                                                                              |=====================================                                 |  53%  |                                                                              |========================================                              |  57%  |                                                                              |==========================================                            |  60%  |                                                                              |============================================                          |  63%  |                                                                              |===============================================                       |  67%  |                                                                              |=================================================                     |  70%  |                                                                              |===================================================                   |  73%  |                                                                              |======================================================                |  77%  |                                                                              |========================================================              |  80%  |                                                                              |==========================================================            |  83%  |                                                                              |=============================================================         |  87%  |                                                                              |===============================================================       |  90%  |                                                                              |=================================================================     |  93%  |                                                                              |====================================================================  |  97%  |                                                                              |======================================================================| 100%

#> ... status: OK

#> [1] TRUE

# Check availability of the corpus
library(polmineR)
corpus() # should include downloaded GERMAPARL(SAMPLE) corpus

#>            corpus   size template
#> 1   GERMAPARLMINI 222201     TRUE
#> 2 GERMAPARLSAMPLE 222201     TRUE
#> 3         REUTERS   4050     TRUE

count("GERMAPARLSAMPLE", "Daten") # an arbitrary test

#>    query count         freq
#> 1: Daten     8 3.600344e-05

germaparl_is_installed(sample = samplemode) # TRUE now

#> [1] TRUE

germaparl_get_version(sample = samplemode) # get version of indexed corpus

#> [1] ‘0.1.0’

germaparl_get_doi(sample = samplemode) # get 'document object identifier' (DOI) of GERMAPARL corpus

#> Warning: Cannot get DOI for corpus GERMAPARL: Corpus has not yet been installed.

#> NULL

# Encode structural attribute 'speech'
germaparl_encode_speeches(
  registry_dir = cwb_dirs[["registry_dir"]],
  corpus_dir = cwb_dirs[["corpus_dir"]],
  sample = samplemode
)

#> ... adding s-attribute 'speech' to registry

#> Corpus to delete (ID): GERMAPARLSAMPLE
#> Corpus name: GermaParlSample: Sample subset of the GermaParl corpus of plenary protocols of the German Bundestag
#> Number of loads before reset: 9
#> Number of loads resetted: 1

#> ... status: OK

# Check whether the new attribute is available 
s_attributes(corpus_id)

#> [1] "interjection" "date"         "party"        "speaker"      "speech"      

values <- s_attributes(corpus_id, "speech")
sizes <- size(corpus_id, s_attribute = "speech")
dtm <- as.DocumentTermMatrix(corpus_id, p_attribute = "word", s_attribute = "speech")

#> ... generate data.table with token and struc ids

#> ... generate unique document ids

#> ... counting token per doc

#> ... generate simple_triplet_matrix

#> ... add row and column labels

# Download topic model (k = 250)
germaparl_download_lda(
  k = 30, # use k = 250 for full GERMAPARL corpus
  data_dir = file.path(cwb_dirs[["corpus_dir"]], tolower(corpus_id)),
  sample = samplemode
)

#> ... downloading: https://zenodo.org/api/files/1f191114-a9fa-4dc9-a4df-950c23b1ff60/germaparlsample_lda_30.rds

lda <- germaparl_load_topicmodel(k = 30L, registry_dir = registry_tmp, sample = samplemode)

#> ... loading topicmodel for k = 30

lda_terms <- topicmodels::terms(lda, 10)

# Encode topic model classification of speeches
germaparl_encode_lda_topics(
  k = 30, # use k = 250 for full GERMAPARL corpus
  n = 3,
  registry_dir = cwb_dirs[["registry_dir"]],
  data_dir = file.path(cwb_dirs[["corpus_dir"]], tolower(corpus_id)),
  sample = samplemode
)

#> ... loading topicmodel for k = 30

#> ... getting topic matrix

#> ... decoding s-attribute speech

#> ... running some sanity checks

#> ... get encoding: latin1

#> ... get cpos and strucs

#> ... encoding s-attribute 'topics'

#> ... adding s-attribute 'topics' to registry

#> Corpus to delete (ID): GERMAPARLSAMPLE
#> Corpus name: GermaParlSample: Sample subset of the GermaParl corpus of plenary protocols of the German Bundestag
#> Number of loads before reset: 33
#> Number of loads resetted: 1

#> ... status: OK

# Check whether the newly encoded attribute 'topics' is available
s_attributes(corpus_id)

#> [1] "interjection" "date"         "party"        "speaker"      "speech"      
#> [6] "topics"      

sc <- corpus(corpus_id) %>% subset(grep("\\|18\\|", topics))
b <- as.speeches(sc, s_attribute_name = "speaker")

#> ... generating partitions by date

#> ... generating speeches

#> ... generating names

#> ... reordering partitions

#> ... coercing partitions to plpr_partitions

length(b)

#> [1] 14