GermaParl R Data Package.

The package offers a convenient dissemination mechanism for the GermaParl corpus that has been linguistically annotated and indexed (data format of the Corpus Workbench / CWB). After installation, only a small sample corpus will be included in the package. Use the germaparl_download_corpus function to download the full corpus. The package offers further functionality to amend the corpus.

germaparl_is_installed()

germaparl_get_doi()

germaparl_get_version()

Details

Note that the R package and the corpus are two different kinds of research data: The package offers a mechanism to ship, easily install and augment the data. The indexed corpus is the actual data. Package and corpus have different version numbers and should be quoted in combination in publications. We recommend to follow the instructions you see when calling citation(package = "GermaParl").

germaparl_is_installed is an auxiliary function that returns TRUE if the corpus has been installed, and FALSE if not.

germaparl_get_doi is an auxiliary function that extracts the DOI (Document Object Identifier) from the registry file of the GERMAPARL corpus. If the corpus has not yet been installed, NULL is returned and a warning will be issued.

germaparl_get_version is an auxiliary function that extracts the version of the GERMAPARL corpus from the registry file. If the corpus has not yet been installed, NULL is returned, and a warning message is issued.

References

Blaette, Andreas (2018): "Using Data Packages to Ship Annotated Corpora of Parliamentary Protocols: The GermaParl R Package". ISBN 979-10-95546-02-3.

Examples

if (FALSE) {
library(polmineR)
use("GermaParl")
corpus() # will include GERMAPARLMINI, sample corpus included in pkg
if (!germaparl_is_installed()) germaparl_download_corpus() # ~1 GB, takes time ...
use("GermaParl")
corpus() # will include GERMAPARL, full corpus
}
germaparl_is_installed() # to check whether GERMAPARL has been downloaded

#> [1] FALSE

germaparl_get_doi() # get 'document object identifier' (DOI) of GERMAPARL corpus

#> Warning: Cannot get DOI for corpus GERMAPARL: Corpus has not yet been installed.

#> NULL

germaparl_get_version

#> function () 
#> {
#>     if (isFALSE(germaparl_is_installed())) {
#>         warning("Cannot get GERMAPARL version: Corpus has not yet been installed.")
#>         return(NULL)
#>     }
#>     regdir <- system.file(package = "GermaParl", "extdata", "cwb", 
#>         "registry")
#>     regdata <- registry_file_parse(corpus = "GERMAPARL", registry_dir = regdir)
#>     regdata[["properties"]][["version"]]
#> }
#> <bytecode: 0x7fdf3e54f540>
#> <environment: namespace:GermaParl>